upokari/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
fsspec==2024.12.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
sympy==1.13.1
torch==2.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0